refactor: move execution instance to new class
